fix png & svg export for no-js
new /export endpoint

basically making the exporting component more lazy

starts off as a form with three radios and a submit button
the form action is clever, if it doesn’t find a hidden value it gets
the data URL from the headers (they don’t count as weight). So that
helps keep the initial load smaller. It doesn’t seemed to have made a
significant enough difference though.

Closes #12